+++ src/main/kotlin/com/knu/mockin/model/dto/kisresponse/quotations/basic/real/KISItemChartPriceResponseDto.kt
@@ -1,13 +1,13 @@
-package com.knu.mockin.model.dto.kisresponse.basic
+package com.knu.mockin.model.dto.kisresponse.quotations.basic.real
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class KISItemChartPriceResponseDto (
-        @JsonProperty("rt_cd") val successFailureStatus: String,    // 성공 실패 여부
-        @JsonProperty("msg_cd") val responseCode: String,           // 응답코드
-        @JsonProperty("msg1") val responseMessage: String,          // 응답메세지
-        @JsonProperty("output1") val output1: ItemChartPriceOutput1?,        // 응답상세
-        @JsonProperty("output2") val output2: List<ItemChartPriceOutput2>?,
+    @JsonProperty("rt_cd") val successFailureStatus: String,    // 성공 실패 여부
+    @JsonProperty("msg_cd") val responseCode: String,           // 응답코드
+    @JsonProperty("msg1") val responseMessage: String,          // 응답메세지
+    @JsonProperty("output1") val output1: ItemChartPriceOutput1?,        // 응답상세
+    @JsonProperty("output2") val output2: List<ItemChartPriceOutput2>?,
 )
 
 data class ItemChartPriceOutput1 (
